publish: add groupify to settings
